remove module name from accept_config*.txt
This is another artifact of supporting separate compilation.

+++ apps/canneal/eval.py
@@ -2,7 +2,7 @@ from __future__ import division
 
 def load(fn='output.txt'):
     with open(fn) as f:
-        return int(f.read())
+        return float(f.read())
 
 def score(orig, relaxed):
     return 1.0 - relaxed / orig
